fix(nvr/syscheck): 32-bit build: cast Totalram to uint64 before multiply
On GOARCH=386, syscall.Sysinfo_t.Totalram is uint32 but was being
multiplied by uint64(info.Unit), causing a type mismatch. Explicit
cast fixes test_32 CI job.

+++ internal/nvr/syscheck/mem_linux.go
@@ -8,5 +8,5 @@ func getSystemMemory() (uint64, error) {
 	if err := syscall.Sysinfo(&info); err != nil {
 		return 0, err
 	}
-	return info.Totalram * uint64(info.Unit), nil
+	return uint64(info.Totalram) * uint64(info.Unit), nil
 }
